+++ package.json
@@ -17,7 +17,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "date-fns": "^2.28.0",
-    "firebase": "^9.6.3",
+    "firebase": "^9.6.4",
     "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
